+++ src/core/reader/types/mod.rs
@@ -11,7 +11,6 @@ use crate::Result;
 use crate::{unreachable_validated, Error};
 
 pub mod data;
-pub mod element;
 pub mod export;
 pub mod function_code_header;
 pub mod global;
